+++ pcs_test/tier0/lib/commands/test_cluster_property.py
@@ -120,6 +120,34 @@ class StonithWatchdogTimeoutMixin(LoadMetadataMixin):
         )
         self.env_assist.assert_reports([])
 
+    def _set_invalid_value(self, forced=False):
+        self.config.remove("services.is_enabled")
+        self.env_assist.assert_raise_library_error(
+            lambda: cluster_property.set_properties(
+                self.env_assist.get_env(),
+                {"stonith-watchdog-timeout": "15x"},
+                [] if not forced else [reports.codes.FORCE],
+            )
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.error(
+                    reports.codes.INVALID_OPTION_VALUE,
+                    option_name="stonith-watchdog-timeout",
+                    option_value="15x",
+                    allowed_values="time interval (e.g. 1, 2s, 3m, 4h, ...)",
+                    cannot_be_empty=False,
+                    forbidden_characters=None,
+                ),
+            ]
+        )
+
+    def test_set_invalid_value(self):
+        self._set_invalid_value(forced=False)
+
+    def test_set_invalid_value_forced(self):
+        self._set_invalid_value(forced=True)
+
 
 class TestSetStonithWatchdogTimeoutSBDIsDisabled(
     StonithWatchdogTimeoutMixin, TestCase
@@ -131,6 +159,9 @@ class TestSetStonithWatchdogTimeoutSBDIsDisabled(
 
     def test_set_zero(self):
         self._set_success({"stonith-watchdog-timeout": "0"})
+
+    def test_set_zero_time_suffix(self):
+        self._set_success({"stonith-watchdog-timeout": "0s"})
 
     def test_set_not_zero_or_empty(self):
         self.env_assist.assert_raise_library_error(
@@ -231,12 +262,12 @@ class TestSetStonithWatchdogTimeoutSBDIsEnabledWatchdogOnly(
     def test_set_zero_forced(self):
         self.config.env.push_cib(
             crm_config=fixture_crm_config_properties(
-                [("cib-bootstrap-options", {"stonith-watchdog-timeout": "0"})]
+                [("cib-bootstrap-options", {"stonith-watchdog-timeout": "0s"})]
             )
         )
         cluster_property.set_properties(
             self.env_assist.get_env(),
-            {"stonith-watchdog-timeout": "0"},
+            {"stonith-watchdog-timeout": "0s"},
             [reports.codes.FORCE],
         )
         self.env_assist.assert_reports(
@@ -271,7 +302,7 @@ class TestSetStonithWatchdogTimeoutSBDIsEnabledWatchdogOnly(
         self.env_assist.assert_raise_library_error(
             lambda: cluster_property.set_properties(
                 self.env_assist.get_env(),
-                {"stonith-watchdog-timeout": "9"},
+                {"stonith-watchdog-timeout": "9s"},
                 [],
             )
         )
@@ -281,7 +312,7 @@ class TestSetStonithWatchdogTimeoutSBDIsEnabledWatchdogOnly(
                     reports.codes.STONITH_WATCHDOG_TIMEOUT_TOO_SMALL,
                     force_code=reports.codes.FORCE,
                     cluster_sbd_watchdog_timeout=10,
-                    entered_watchdog_timeout="9",
+                    entered_watchdog_timeout="9s",
                 )
             ]
         )
@@ -289,12 +320,12 @@ class TestSetStonithWatchdogTimeoutSBDIsEnabledWatchdogOnly(
     def test_too_small_forced(self):
         self.config.env.push_cib(
             crm_config=fixture_crm_config_properties(
-                [("cib-bootstrap-options", {"stonith-watchdog-timeout": "9"})]
+                [("cib-bootstrap-options", {"stonith-watchdog-timeout": "9s"})]
             )
         )
         cluster_property.set_properties(
             self.env_assist.get_env(),
-            {"stonith-watchdog-timeout": "9"},
+            {"stonith-watchdog-timeout": "9s"},
             [reports.codes.FORCE],
         )
         self.env_assist.assert_reports(
@@ -302,13 +333,13 @@ class TestSetStonithWatchdogTimeoutSBDIsEnabledWatchdogOnly(
                 fixture.warn(
                     reports.codes.STONITH_WATCHDOG_TIMEOUT_TOO_SMALL,
                     cluster_sbd_watchdog_timeout=10,
-                    entered_watchdog_timeout="9",
+                    entered_watchdog_timeout="9s",
                 )
             ]
         )
 
     def test_more_than_timeout(self):
-        self._set_success({"stonith-watchdog-timeout": "11"})
+        self._set_success({"stonith-watchdog-timeout": "11s"})
 
 
 @mock.patch("pcs.lib.sbd.get_local_sbd_device_list", lambda: ["dev1", "dev2"])
@@ -322,6 +353,9 @@ class TestSetStonithWatchdogTimeoutSBDIsEnabledSharedDevices(
 
     def test_set_to_zero(self):
         self._set_success({"stonith-watchdog-timeout": "0"})
+
+    def test_set_to_zero_time_suffix(self):
+        self._set_success({"stonith-watchdog-timeout": "0min"})
 
     def test_set_not_zero_or_empty(self):
         self.env_assist.assert_raise_library_error(