+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 - Displaying bool and integer values in `pcs resource config` command
   ([rhbz#2151166], [ghissue#604])
+- Allow time values in stonith-watchdog-time property ([rhbz#2158804])
 
 ### Changed
 - Resource/stonith agent self-validation of instance attributes is now
@@ -20,6 +21,7 @@
 [ghissue#604]: https://github.com/ClusterLabs/pcs/issues/604
 [rhbz#2151166]: https://bugzilla.redhat.com/show_bug.cgi?id=2151166
 [rhbz#2151511]: https://bugzilla.redhat.com/show_bug.cgi?id=2151511
+[rhbz#2158804]: https://bugzilla.redhat.com/show_bug.cgi?id=2158804
 [rhbz#2159455]: https://bugzilla.redhat.com/show_bug.cgi?id=2159455